+++ madeasy/parser/views.py
@@ -1,6 +1,7 @@
 import datetime
 
 from django.db.models import Avg
+from django.utils import timezone
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
@@ -29,6 +30,7 @@ class ParserView(APIView):
         parser_results = ParserResults()
         query_raw = request.data.get('query', '')
         parser_results.query_raw = query_raw
+        parser_results.created_by = request.user
         parser_results.save()
         if serializer.is_valid():
             query = serializer.data.get('query')
@@ -53,7 +55,6 @@ class ParserView(APIView):
                 parser_results.save()
                 return Response(error, status=status.HTTP_400_BAD_REQUEST)
         parser_results.save()
-        print(request.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -68,6 +69,33 @@ class ParserResultsViewSet(viewsets.ModelViewSet):
         neg = ParserResults.objects.filter(is_correct=False)
 
         parser_correctness = {
+            "positive": pos.count(),
+            "negative": neg.count()
+        }
+        return Response(parser_correctness, status=status.HTTP_200_OK)
+
+    @list_route(methods=['get'])
+    def parser_command_type(self, request):
+        unknown = ParserResults.objects.filter(command_executed="Unkwown")
+        book = ParserResults.objects.filter(command_executed="BookCommand")
+        show = ParserResults.objects.filter(command_executed="ShowCommand")
+
+        parser_correctness = {
+            "unknown": unknown.count(),
+            "book": book.count(),
+            "show": show.count()
+        }
+        return Response(parser_correctness, status=status.HTTP_200_OK)
+
+    @list_route(methods=['get'])
+    def parser_numbers(self, request):
+        today = ParserResults.objects.filter(
+            created__gt=timezone.now().date())
+        pos = ParserResults.objects.filter(is_correct=True)
+        neg = ParserResults.objects.filter(is_correct=False)
+
+        parser_correctness = {
+            "today": today.count(),
             "positive": pos.count(),
             "negative": neg.count()
         }