+++ CHANGELOG.md
@@ -82,3 +82,10 @@ CHANGELOG
 ## 0.0.16 - 2016-05-17
 - Add show my booked flights command
 - Edit book command to allow search by price and to specify a return date
+
+## 0.0.16 - 2016-05-18
+- Add actions parameter to the user fields
+
+## 0.0.17 - 2016-05-20
+- Add command for adults
+- Special endpoints for parser results